minor changes to carousel format. Added in action image on front page. moved in some content for about and started formatting login page
