+++ client/src/components/About.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const About = () => {
+    return (
+        <div>
+            <p>"Every year, like clockwork, DJ Bentley is a regular presence at the CWMAs, and for good reason—he’s a goddammed killer on the decks. Few DJs have the ability to flawlessly bounce around through genres as this Salt Lake City-based DJ does. Bentley is a resident DJ at Park City Live and has been a fixture at many of Utah’s biggest parties, like the Greek Knights New Year’s Eve Gala. Expect Bentley to play a healthy mix of rap, pop, trap, house, whatever—he can do it all." - City Weekly</p>
+
+            <p>DJ Bentley has all the skills needed to bridge the gap between any genre. He keeps party-goers on their toes non-stop with his vast and intricate mix of dance, house, pop, hip-hop and more. Most recently, he has become a weekly resident DJ at Salt Lake City’s hottest night club The Moose Lounge each and every Saturday night along side another of Utah’s premiere DJ’s Luva Luva. You can also find DJ Bentley cutting records each and every year at The Sundance Film Festival as well as many of the other major private and corporate events Utah has to offer.</p>
+        </div>
+    )
+}
+
+export default About
