+++ client/src/components/MusicPlayer.js
