+++ client/src/style/_pages.scss
@@ -3,10 +3,16 @@
 .btn {
     @include pad-mar(8px, 10px);
     border-radius: 5px;
-    border: 1px solid $accentColor;
+    border: 1px solid $secondaryAccentColor;
     background-color: $darkColor;
     color: $lightColor;
     font-size: 20px;
+}
+
+.btn:hover {
+    box-shadow: $secondaryAccentColor 0px 0px 10px;
+    color: $accentColor;
+    background-color: $transparentDarkColor;
 }
 
 .sideLink {
@@ -33,6 +39,7 @@
     width: 50px;
     height: 50px;
     background-color: $darkColor;
+    box-shadow: $secondaryAccentColor 0px 0px 10px;
 
     svg {
         color: $accentColor;
@@ -92,8 +99,10 @@
     border: black solid 1px;
     width: 100%;
     height: 75vh;
-    background-color: $transparentDarkColor;
-    color: $lightColor;
+    // background-color: $transparentDarkColor;
+    // color: $lightColor;
+    background-image: url('../assets/ActionImage_DJBentley.jpg');
+    background-size: 100%;
 }
 
 .carousel {
